scripts : allow to exec scripts only from main/root directory

+++ scripts/build_musl.sh
@@ -2,6 +2,11 @@
 
 NJOBS=`nproc`
 
+if [[ `basename $PWD` != "nulix" ]]; then
+	echo "This script must be run from main/root directory"
+	exit 1
+fi
+
 # create a build directory
 cd musl
 rm -rf musl-build
